refactor: `modules/float` is now `modules/pool`
