+++ src/index.ts
@@ -13,7 +13,7 @@ export { buildContext } from './modules/context';
 export { isAddressValid } from './modules/validate-address';
 export { buildExplorerLink } from './modules/explorer-link';
 export { getChainFor, isSkybridgeChain } from './modules/chains';
-export { createFloat } from './modules/float';
+export { createFloat } from './modules/pool';
 export { isSkybridgeBridge } from './modules/bridges';
 export { isSkybridgeMode } from './modules/modes';
 export { isSkybridgeResource } from './modules/resources';
